Use github_issue_label instead of github_issue_labels
The github_issue_labels module deals poorly with the labels that github
creates by default in new repositories. When attempting to apply a new
github_issue_labels resource, this will invariably result in a conflict
error, and require a manual resource import to rectify.

We can avoid this by using the github_issue_label module, but with the
disadvantage that we are now managing individual labels, rather than an
exclusive set. In particular, that means we won't remove labels that were
created manually. That's probably okay.