Disable automerge on private repositories
It turns out that GitHub doesn't support auto-merging pull requests on
private repositories [1]. The request to set the attribute is accepted
without error, but doesn't actually change anything. This causes noise in
the output of `tofu apply` since it's constantly trying to re-apply the
change.

Update the `common_repository` module so that we set
`allow_auto_merge=false` for private repositories.

[1]: https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/incorporating-changes-from-a-pull-request/automatically-merging-a-pull-request